Delete unused config file
